change-valid-email

+++ src/public/javascripts/main.js
@@ -42,7 +42,7 @@ function scrollLoadData() {
   let limit = 10;
   if (check.length >= 21 && path === 'profile' || check === "") {
     $(window).scroll( () => {
-      if (Math.abs($(window).scrollTop() - ($(document).height() - $(window).height())) < 1 ) {
+      if (Math.abs($(window).scrollTop() - ($(document).height() - $(window).height())) < 2 ) {
         fetch("/post/load", {
           method: "POST",
           headers: { "Content-Type": "application/json" },
@@ -60,6 +60,7 @@ function scrollLoadData() {
                 let id = $("#index_id_user").html();
                 start = start + 1;
                 let p = json.post;
+                let user = json.user
                 let comments = json.comments;
                 p.forEach((post) => {
                   if (id === post.user.id) {
@@ -99,7 +100,7 @@ function scrollLoadData() {
                     }
                   } else {
                     if (post.urlFile.length > 0) {
-                      $(".index_body_post").append(renderPostImageDifferentUser(post));
+                      $(".index_body_post").append(renderPostImageDifferentUser(post,user));
                       comments.forEach((cmt) => {
                         if (cmt.idPost === post.id) {
                           if(cmt.user.id === id){
@@ -110,7 +111,7 @@ function scrollLoadData() {
                         }
                       });
                     } else if (post.idVideos.length > 0) {
-                      $(".index_body_post").append(renderPostVideoDifferentUser(post));
+                      $(".index_body_post").append(renderPostVideoDifferentUser(post,user));
                       comments.forEach((cmt) => {
                         if (cmt.idPost === post.id) {
                           if(cmt.user.id === id){
@@ -121,7 +122,7 @@ function scrollLoadData() {
                         }
                       });
                     } else {
-                      $(".index_body_post").append(renderPostCharDifferentUser(post))
+                      $(".index_body_post").append(renderPostCharDifferentUser(post,user))
                       comments.forEach((cmt) => {
                         if (cmt.idPost === post.id) {
                           if(cmt.user.id === id){
@@ -346,18 +347,16 @@ function addPost() {
         if (xhr.readyState === 4 && xhr.status === 200) {
           let json = JSON.parse(xhr.responseText);
           if (json.code === 0) {
+            $("#index_modal_new_post").modal("hide");
             let post = json.post;
             if (post.urlFile.length > 0) {
               $(".index_body_post").prepend(renderPostImage(post))
-              $("#index_modal_new_post").modal("hide");
               renderAlert()
             } else if (post.idVideos.length > 0) {
               $(".index_body_post").prepend(renderPostVideo(post))
-              $("#index_modal_new_post").modal("hide");
               renderAlert()
             } else {
               $(".index_body_post").prepend(renderPostChar(post))
-              $("#index_modal_new_post").modal("hide");
               renderAlert()
             }
           } else if (json.code === 1) {
@@ -505,17 +504,19 @@ function editPost(e) {
 
 function deletePost(e) {
   let id = e.id;
-  fetch("/post/delete/" + id, {
-    method: "DELETE",
-  })
-    .then((res) => res.json())
-    .then((json) => {
-      if (json.code === 0) {
-        $(`div#${id}`).remove();
-        renderAlert()
-      }
+  if(confirm("Are you sure you want to delete this post ?")) {
+    fetch("/post/delete/" + id, {
+      method: "DELETE",
     })
-    .catch((err) => console.log(err));
+      .then((res) => res.json())
+      .then((json) => {
+        if (json.code === 0) {
+          $(`div#${id}`).remove();
+          renderAlert()
+        }
+      })
+      .catch((err) => console.log(err));
+  }
 }
 
 function addComments(e) {
@@ -545,18 +546,20 @@ function addComments(e) {
 
 function deleteComments(e) {
   let id = e.id;
-  fetch("/comments/delete/" + id, {
-    method: "DELETE",
-  })
-    .then((res) => res.json())
-    .then((json) => {
-      if (json.code === 0) {
-        socket.emit("client-send-delete-comment",id)
-        $(`div#${id}`).remove();
-        renderAlert()
-      }
+  if(confirm('Are you sure you want to delete this comment')) {
+    fetch("/comments/delete/" + id, {
+      method: "DELETE",
     })
-    .catch((err) => console.log(err));
+      .then((res) => res.json())
+      .then((json) => {
+        if (json.code === 0) {
+          socket.emit("client-send-delete-comment",id)
+          $(`div#${id}`).remove();
+          renderAlert()
+        }
+      })
+      .catch((err) => console.log(err));
+  }
 }
 
 function addNotification() {
@@ -611,17 +614,19 @@ function addNotification() {
 
 function deleteNotification(e) {
     let id = e.id
-    fetch("/notification/delete/" + id, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((json) => {
-        if (json.code === 0) {
-          $(`li#${id}`).remove();
-          renderAlert()
-        }
+    if(confirm('Are you sure you want to delete this notification')) {
+      fetch("/notification/delete/" + id, {
+        method: "DELETE",
       })
-      .catch((err) => console.log(err));
+        .then((res) => res.json())
+        .then((json) => {
+          if (json.code === 0) {
+            $(`li#${id}`).remove();
+            renderAlert()
+          }
+        })
+        .catch((err) => console.log(err));
+    }
 }
 
 function editNotification(e) {
@@ -774,7 +779,7 @@ function renderPostChar(post){
     </div>`
 }
 
-function renderPostCharDifferentUser(post){ 
+function renderPostCharDifferentUser(post,user){ 
   let time = moment(post.time).fromNow()
   return `<div class="card index_poster" id="${post.id}">
     <div class="card index_on_the_cmt">
@@ -786,7 +791,7 @@ function renderPostCharDifferentUser(post){
       <button type="button" class="btn btn-light index_button_cmt" id="" name="" value="" data-toggle="collapse" href="#collapseCmt${post.id}" role="button" aria-expanded="false" aria-controls="collapseCmt${post.id}"><i class="far fa-comment-dots"></i>  Comment</button>
       <div class="collapse" id="collapseCmt${post.id}">
         <div class=" index_cmt" id="${post.id}>">
-          <img id="profile_avt" class="index_avt_cmt" src="${post.user.img}">
+          <img id="profile_avt" class="index_avt_cmt" src="${user.img}">
           <input type="text" class="form-control index_cmt_texbox" id="index_data_cmt${post.id}" value="" placeholder="Your comment..." name="">
           <button id="${post.id}" class="btn btn-primary" onclick="addComments(this)">Send</button>
          </div>
@@ -825,7 +830,7 @@ function renderPostVideo(post){
   </div>`
 }
 
-function renderPostVideoDifferentUser(post){
+function renderPostVideoDifferentUser(post,user){
   let time = moment(post.time).fromNow()
   return `<div class="card index_poster" id="${post.id}">
     <div class="card index_on_the_cmt">
@@ -839,7 +844,7 @@ function renderPostVideoDifferentUser(post){
     <button type="button" class="btn btn-light index_button_cmt" id="" name="" value="" data-toggle="collapse" href="#collapseCmt${post.id}" role="button" aria-expanded="false" aria-controls="collapseCmt${post.id}"><i class="far fa-comment-dots"></i>  Comment</button>
     <div class="collapse" id="collapseCmt${post.id}">
       <div class=" index_cmt" id="${post.id}>">
-        <img id="profile_avt" class="index_avt_cmt" src="${post.user.img}">
+        <img id="profile_avt" class="index_avt_cmt" src="${user.img}">
         <input type="text" class="form-control index_cmt_texbox" id="index_data_cmt${post.id}" value="" placeholder="Your comment..." name="">
         <button id="${post.id}" class="btn btn-primary" onclick="addComments(this)">Send</button>
       </div>
@@ -878,7 +883,7 @@ function renderPostImage(post){
     </div>`
 }
 
-function renderPostImageDifferentUser(post){
+function renderPostImageDifferentUser(post,user){
   let time = moment(post.time).fromNow()
    return `<div class="card index_poster" id="${post.id}">
       <div class="card index_on_the_cmt">
@@ -892,7 +897,7 @@ function renderPostImageDifferentUser(post){
         <button type="button" class="btn btn-light index_button_cmt" id="" name="" value="" data-toggle="collapse" href="#collapseCmt${post.id}" role="button" aria-expanded="false" aria-controls="collapseCmt${post.id}"><i class="far fa-comment-dots"></i>  Comment</button>
         <div class="collapse" id="collapseCmt${post.id}">
           <div class="index_cmt" id="${post.id}>">
-            <img id="profile_avt" class="index_avt_cmt" src="${post.user.img}">
+            <img id="profile_avt" class="index_avt_cmt" src="${user.img}">
             <input type="text" class="form-control index_cmt_texbox" id="index_data_cmt${post.id}" value="" placeholder="Your comment..." name="">
             <button id="${post.id}" class="btn btn-primary" onclick="addComments(this)">Send</button>
           </div>
